requirements.txt
asgiref==3.8.1
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
Django==5.1.4
docx2pdf==0.1.8
img2pdf==0.5.1
lxml==5.3.0
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pikepdf==9.5.1
pillow==11.1.0
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.1
python-pptx==1.0.2
pywin32==308
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
wrapt==1.17.0
XlsxWriter==3.2.0
